moving fskeystore to new file
